misc/ac_2019/requirements.txt
appdirs==1.4.3
astroid==2.3.3
beautifulsoup4==4.8.1
black==19.10b0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
fuse-python==1.0.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.1
mccabe==0.6.1
online-judge-tools==7.3.0
pathspec==0.6.0
pylint==2.4.4
regex==2019.11.1
requests==2.22.0
requests-cache==0.5.2
rope==0.14.0
soupsieve==1.9.5
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
